Move sorting flag and mode in dca namespace and shorten mode names
