+++ library/Haste/Dca/SortingMode.php
@@ -10,7 +10,7 @@
  * @license    http://opensource.org/licenses/lgpl-3.0.html LGPL
  */
 
-namespace Haste\Util;
+namespace Haste\Dca;
 
 
 /**
@@ -25,36 +25,36 @@ class SortingMode
     /**
      * Records are not sorted
      */
-    const NOT_SORTED = 0;
+    const NONE = 0;
 
     /**
      * Records are sorted by a fixed field
      */
-    const SORTED_FIXED_FIELD = 1;
+    const FIXED_FIELD = 1;
 
     /**
      * Records are sorted by a switchable field
      */
-    const SORTED_SWITCHABLE_FIELD = 2;
+    const SWITCHABLE_FIELD = 2;
 
     /**
      * Records are sorted by the parent table
      */
-    const SORTED_PARENT_TABLE = 3;
+    const PARENT_TABLE = 3;
 
     /**
      * Displays the child records of a parent record (see style sheets module)
      */
-    const DISPLAY_AS_CHILD = 4;
+    const CHILD_VIEW = 4;
 
     /**
      * Records are displayed as tree (see site structure)
      */
-    const DISPLAY_AS_TREE = 5;
+    const TREE_VIEW = 5;
 
     /**
      * Displays the child records within a tree structure (see articles module)
      */
-    const DISPLAY_AS_CHILD_TREE = 6;
+    const CHILD_TREE_VIEW = 6;
 
 }
